+++ CHANGELOG.md
@@ -15,7 +15,9 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Fixed
 
-- None
+- [#1084](https://github.com/paper-trail-gem/paper_trail/pull/1084)
+  The `touch` callback (added in 9.0.0) now inserts the correct value into
+  the `versions.object` column.
 
 ### Other